Refactor Heroic Green Steel components and crafting logic
- Remove `InvasionList.tsx` to replace functionality with new approach.
- Add `MaterialsAccordion.tsx` for crafting material categorization.
- Replace inline Accordion logic with `MaterialsAccordion` in `IngredientList.tsx`.
- Update `Countdown.tsx` with stricter typing using `luxon` types.
- Simplify action reducers in `hgsSlice.ts` by consolidating logic.
- Remove dead code and unused imports in `IngredientList.tsx` and slices.
