+++ src/components/timer/Countdown.tsx
@@ -9,34 +9,32 @@ const Countdown = (props: Props) => {
 
   useEffect(() => {
     const updateRemainingTime = () => {
-      const now = DateTime.now()
-      const targetTime = DateTime.fromSeconds(targetTimestamp)
-      const diff = targetTime.diff(now, ['days', 'hours', 'minutes', 'seconds'])
+      const now: DateTime<Valid> = DateTime.now()
+      const targetTime: DateTime<Valid> = DateTime.fromSeconds(targetTimestamp)
+      const diff: Duration<Valid> = targetTime.diff(now, ['days', 'hours', 'minutes', 'seconds'])
 
       if (diff.toMillis() <= 0) {
-        setRemainingTime(Duration.fromMillis(0)) // Stop countdown at 0
-        if (onComplete) onComplete() // Call onComplete callback (if any)
+        setRemainingTime(Duration.fromMillis(0))
+        if (onComplete) onComplete()
         return
       }
 
       setRemainingTime(diff)
     }
 
-    // Initial calculation
     updateRemainingTime()
 
-    // Update every second
     const interval = setInterval(updateRemainingTime, 1000)
 
     return () => {
       clearInterval(interval)
-    } // Clear interval on unmount
+    }
   }, [onComplete, targetTimestamp])
 
-  const days = remainingTime.days || 0
-  const hours = remainingTime.hours || 0
-  const minutes = remainingTime.minutes || 0
-  const seconds = Math.floor(remainingTime.seconds || 0)
+  const days: number = remainingTime.days || 0
+  const hours: number = remainingTime.hours || 0
+  const minutes: number = remainingTime.minutes || 0
+  const seconds: number = Math.floor(remainingTime.seconds || 0)
 
   return (
     <span>
